+++ src/main/java/livenessanalysis/DataflowGraph.java
@@ -1,6 +1,7 @@
 package livenessanalysis;
 
 import analysis.DepthFirstAdapter;
+import java.util.HashSet;
 import java.util.PriorityQueue;
 import lineevaluation.LineEvaluator;
 import node.AAssignStat;
@@ -21,31 +22,38 @@ import symboltable.Symbol;
 import symboltable.SymbolTable;
 
 /**
- * Builder for a dataflow graph. Performs a backwards analysis of the AST (i.e.,
- * uses a depth first approach). To build a dataflow graph, apply this class to
- * a function-related sub-tree of the AST.
+ * Dataflow graph with one node per statement. Nodes register the symbol
+ * definition and usage across the program.
  */
-public class DataflowGraphBuilder extends DepthFirstAdapter {
+public class DataflowGraph extends DepthFirstAdapter {
   private final SymbolTable symbolTable;
   private final LineEvaluator lineEvaluator;
   private DataflowNode start;
+  private DataflowNode end;
   private DataflowNode current;
   private int currentNumber;
 
   /**
-   * Builder for a dataflow graph. Performs a backwards analysis of the AST (i.e.,
-   * uses a depth first approach). To build a dataflow graph, apply this class to
-   * a function-related sub-tree of the AST.
+   * Builder for a dataflow graph. Walks the AST to create a node per statement.
+   * Performs a backwards analysis of the resulting graph to determine in- and
+   * out-sets.
    *
    * @param symbolTable   Filled symbol table for the corresponding AST.
    * @param lineEvaluator Line evaluator that has been applied to the AST.
+   * @param astStart      Start node of the (function-related sub-)AST.
    */
-  public DataflowGraphBuilder(SymbolTable symbolTable, LineEvaluator lineEvaluator) {
+  public DataflowGraph(
+      SymbolTable symbolTable, LineEvaluator lineEvaluator, Node astStart) {
     this.symbolTable = symbolTable;
     this.lineEvaluator = lineEvaluator;
     this.start = null;
+    this.end = null;
     this.currentNumber = 0;
     this.current = null;
+
+    astStart.apply(this);
+    this.end = this.current;
+    generateInAndOutSets();
   }
 
   // Function heads: define dataflow graph start node
@@ -80,8 +88,8 @@ public class DataflowGraphBuilder extends DepthFirstAdapter {
     DataflowNode successor = getNewSymbolNode(node);
     successor.addDef(defSymbol);
 
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
   }
 
   @Override
@@ -93,37 +101,37 @@ public class DataflowGraphBuilder extends DepthFirstAdapter {
     DataflowNode successor = getNewSymbolNode(node);
     successor.addDef(defSymbol);
 
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
   }
 
   // Possibly reading statements: create new nodes
   @Override
   public void inADeclStat(ADeclStat node) {
     DataflowNode successor = getNewSymbolNode(node);
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
   }
 
   @Override
   public void inAPrintStat(APrintStat node) {
     DataflowNode successor = getNewSymbolNode(node);
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
   }
 
   @Override
   public void inAPrintlnStat(APrintlnStat node) {
     DataflowNode successor = getNewSymbolNode(node);
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
   }
 
   @Override
   public void caseAIfStat(AIfStat node) {
     DataflowNode successor = getNewSymbolNode(node);
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
 
     node.getExpr().apply(this);
     node.getThenBlock().apply(this);
@@ -132,8 +140,8 @@ public class DataflowGraphBuilder extends DepthFirstAdapter {
   @Override
   public void inAIfelseStat(AIfelseStat node) {
     DataflowNode successor = getNewSymbolNode(node);
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
 
     node.getExpr().apply(this);
     node.getThenBlock().apply(this);
@@ -143,22 +151,22 @@ public class DataflowGraphBuilder extends DepthFirstAdapter {
   @Override
   public void inAWhileStat(AWhileStat node) {
     DataflowNode successor = getNewSymbolNode(node);
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
 
     node.getExpr().apply(this);
     node.getBody().apply(this);
 
     // current is the last statement of while-body now and needs to point on the
     // head of this while loop
-    current.addEdgeTo(successor);
+    this.current.addEdgeTo(successor);
   }
 
   @Override
   public void inAReturnStat(AReturnStat node) {
     DataflowNode successor = getNewSymbolNode(node);
-    current.addEdgeTo(successor);
-    current = successor;
+    this.current.addEdgeTo(successor);
+    this.current = successor;
   }
 
   // Expressions: add symbol to current DataflowNode's use-set ("read" statement)
@@ -167,7 +175,7 @@ public class DataflowGraphBuilder extends DepthFirstAdapter {
     String id = node.getId().getText();
     String scopeName = symbolTable.determineScope(node);
     Symbol symbol = symbolTable.getSymbol(scopeName, id);
-    current.addUse(symbol);
+    this.current.addUse(symbol);
   }
 
   // Helpers
@@ -181,8 +189,8 @@ public class DataflowGraphBuilder extends DepthFirstAdapter {
     return this.start;
   }
 
-  DataflowNode getCurrent() {
-    return this.current;
+  DataflowNode getEnd() {
+    return this.end;
   }
 
   /** Print the resulting dataflow graph including relevant sets. */
@@ -203,5 +211,68 @@ public class DataflowGraphBuilder extends DepthFirstAdapter {
       }
     }
     System.out.println();
+  }
+
+  /*
+   * Analyze the dataflow graph using Algorithm 10.4 from Appel, "Modern Compiler
+   * Impl. in Java" to determine in- and out-sets. Initial in- and out-sets are
+   * empty. The sets can be used to determine unused variables.
+   */
+  void generateInAndOutSets() {
+    if (this.end.getPredecessors().isEmpty()) {
+      return;
+    }
+    boolean changes = true;
+    PriorityQueue<DataflowNode> queue = new PriorityQueue<>();
+
+    while (changes) {
+      changes = false;
+
+      DataflowNode curr;
+      queue.add(this.end);
+      while (queue.peek() != null) {
+        curr = queue.poll();
+        int currNumber = curr.getNumber();
+
+        HashSet<Symbol> oldIn = new HashSet<>(curr.getIn());
+        HashSet<Symbol> oldOut = new HashSet<>(curr.getOut());
+        curr.addIn(getUseJointOutMinusDef(curr));
+        curr.addOut(getJoinedSuccessorIns(curr));
+
+        if (differentElementsIn(oldIn, curr.getIn())
+            || differentElementsIn(oldOut, curr.getOut())) {
+          changes = true;
+        }
+        for (DataflowNode predecessor : curr.getPredecessors()) {
+          if (predecessor.getNumber() < currNumber) {
+            queue.add(predecessor);
+          }
+        }
+      }
+    }
+  }
+
+  /* use[n] joint with (out[n]-def[n]). */
+  private HashSet<Symbol> getUseJointOutMinusDef(DataflowNode node) {
+    HashSet<Symbol> returnSet = new HashSet<>(node.getUse());
+
+    HashSet<Symbol> difference = new HashSet<>(node.getOut());
+    difference.removeAll(node.getDef());
+    returnSet.addAll(difference);
+
+    return returnSet;
+  }
+
+  private HashSet<Symbol> getJoinedSuccessorIns(DataflowNode node) {
+    HashSet<Symbol> returnSet = new HashSet<>();
+    HashSet<DataflowNode> successors = node.getSuccessors();
+
+    successors.forEach(successor -> returnSet.addAll(successor.getIn()));
+
+    return returnSet;
+  }
+
+  private boolean differentElementsIn(HashSet<Symbol> setA, HashSet<Symbol> setB) {
+    return !setA.containsAll(setB) || !setB.containsAll(setA);
   }
 }